+++ app/src/main/java/agro/meteoro/wayanad/Initiate.java
@@ -69,7 +69,7 @@ public class Initiate extends AppCompatActivity {
         }
         else
         {
-            startActivity(new Intent(Initiate.this, Preferences.class));
+            startActivity(new Intent(Initiate.this, SelectLanguage.class));
             finish();
         }
     }
@@ -84,7 +84,7 @@ public class Initiate extends AppCompatActivity {
                 {
                     if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED)
                     {
-                        startActivity(new Intent(Initiate.this, Preferences.class));
+                        startActivity(new Intent(Initiate.this, SelectLanguage.class));
                         finish();
                     }
                 }
