+++ app/src/main/java/agro/meteoro/wayanad/SelectCrop.java
@@ -1,0 +1,14 @@
+package agro.meteoro.wayanad;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class SelectCrop extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_select_crop);
+    }
+}
