+++ app/src/main/res/layout/activity_select_crop.xml
@@ -1,79 +1,80 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/post_splash"
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center_horizontal"
-    android:background="@color/bg_2"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
+    tools:context=".SelectCrop">
     <TextView
         android:id="@+id/splash_title_2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginTop="120dp"
+        android:fontFamily="sans-serif-light"
+        android:text="@string/crop_title"
         android:textAlignment="center"
+        android:textColor="@android:color/white"
         android:textSize="30sp"
         android:textStyle="bold"
-        android:layout_marginTop="120dp"
-        android:textColor="@android:color/white"
-        android:fontFamily="sans-serif-light"
-        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:text="@string/crop_title"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
     <LinearLayout
-        android:id="@+id/linear_radios"
+        android:id="@+id/linear_checks"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:gravity="center|left"
-        app:layout_constraintStart_toStartOf="parent"
+        android:gravity="left|center"
+        android:orientation="vertical"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/splash_title_2"
-        android:orientation="vertical">
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
         <CheckBox
-            android:id="@+id/rice_check"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/crop_rice"
-            android:textSize="20sp"
-            android:textColor="@android:color/white"
-            android:buttonTint="@android:color/white"/>
-        <CheckBox
-            android:id="@+id/tea_check"
+            android:id="@+id/check_tea"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/crop_tea"
-            android:textSize="20sp"
+            android:buttonTint="@android:color/white"
             android:textColor="@android:color/white"
-            android:buttonTint="@android:color/white"/>
+            android:textSize="20sp"/>
         <CheckBox
-            android:id="@+id/coffee_check"
+            android:id="@+id/check_rice"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/crop_rice"
+            android:buttonTint="@android:color/white"
+            android:textColor="@android:color/white"
+            android:textSize="20sp"/>
+        <CheckBox
+            android:id="@+id/check_coffee"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/crop_coffee"
-            android:textSize="20sp"
+            android:buttonTint="@android:color/white"
             android:textColor="@android:color/white"
-            android:buttonTint="@android:color/white"/>
+            android:textSize="20sp"/>
         <CheckBox
-            android:id="@+id/bpepper_check"
+            android:id="@+id/check_bpepper"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/crop_bpepper"
-            android:textSize="20sp"
+            android:buttonTint="@android:color/white"
             android:textColor="@android:color/white"
-            android:buttonTint="@android:color/white"/>
+            android:textSize="20sp"/>
+
     </LinearLayout>
     <ImageButton
         android:id="@+id/next_2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:background="@drawable/circle_shape"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
         android:layout_marginBottom="30dp"
+        android:background="@drawable/circle_shape"
+        android:src="@drawable/ic_next"
         android:tint="@android:color/white"
-        android:src="@drawable/ic_next"/>
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
